[BE] feat #115: 이력서 업로드 테스트

+++ backend/d102-file/src/main/java/com/d102/file/dto/UploadDto.java
@@ -22,7 +22,7 @@ public class UploadDto {
     @Data
     public static class ResumeRequest {
 
-        private int beforeId;
+        private Long beforeId;
         private MultipartFile resume;
     }
